ログイン周りの変更とログインチェック関数の作成
