+++ python/app/templates/login.html
@@ -13,6 +13,7 @@
         <p class="text-center text-red-500">{{ error_message }}</p>
         <h2 class="text-2xl font-bold text-center mb-6">ログイン</h2>
         <form action="/login" method="POST">
+            <input type="hidden" id="url" name="url" value="{{ url }}">
             <div class="mb-6">
                 <label for="email" class="block text-gray-700 text-sm font-bold mb-2">メールアドレス</label>
                 <input type="email" id="email" name="email" required
